chore: moved scripts to index ionstead of app
