+++ pages/index.tsx
@@ -23,6 +23,7 @@ import SliderPortalverseWrapper from "@/components/SliderPortalverseWrapper"
 import OfertaEducativaWrapper from "@/components/OfertaEducativaWrapper"
 import NumbersPortalverseWrapper from "@/components/NumbersPortalverseWrapper"
 import BlogPostCardWrapper from "@/components/BlogPostCardWrapper"
+import Script from "next/script"
 
 const Home: NextPageWithLayout = ({ data: { sections, meta, strapi } }: any) => {
   const router = useRouter();
@@ -58,6 +59,69 @@ const Home: NextPageWithLayout = ({ data: { sections, meta, strapi } }: any) => 
 
   return <>
     <Head>
+      <Script  src="https://www.googletagmanager.com/gtag/js?id=G-1ZX45TXRD5" />
+      <Script id='google-analytics'
+        dangerouslySetInnerHTML={{
+          __html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', 'G-1ZX45TXRD5', {
+              page_path: window.location.pathname,
+            });
+          `,
+      }}/>
+      <Script id='google-tag-manager'
+      dangerouslySetInnerHTML={{
+        __html: `(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
+        new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
+        j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
+        'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
+        })(window,document,'script','dataLayer','GTM-NGBCMHM');`,
+      }}/>
+      <Script id='hot-jar'
+        dangerouslySetInnerHTML={{
+          __html: `(function(h,o,t,j,a,r){
+            h.hj=h.hj||function(){(h.hj.q=h.hj.q||[]).push(arguments)};
+            h._hjSettings={hjid:3340067,hjsv:6};
+            a=o.getElementsByTagName('head')[0];
+            r=o.createElement('script');r.async=1;
+            r.src=t+h._hjSettings.hjid+j+h._hjSettings.hjsv;
+            a.appendChild(r);
+            })(window,document,'https://static.hotjar.com/c/hotjar-','.js?sv=');`,
+      }}/>
+      <Script id='meta-pixel'
+        dangerouslySetInnerHTML={{
+          __html: `
+          !function(f,b,e,v,n,t,s){if(f.fbq)return;n=f.fbq=function(){n.callMethod?n.callMethod.apply(n,arguments):n.queue.push(arguments)};
+          if(!f._fbq)f._fbq=n;
+          n.push=n;
+          n.loaded=!0;
+          n.version='2.0';
+          n.queue=[];
+          t=b.createElement(e);
+          t.async=!0;
+          t.src=v;
+          s=b.getElementsByTagName(e)[0];
+          s.parentNode.insertBefore(t,s)}(window, document,'script','https://connect.facebook.net/en_US/fbevents.js');
+          fbq('init', '487461332128996');
+          fbq('track', 'PageView');
+          `,
+      }}/>
+      <Script id='meta-pixel-registration'
+        dangerouslySetInnerHTML={{
+          __html: `
+          fbq('track', 'CompleteRegistration');          
+          `,
+      }}/>
+      {/* // // Meta Pixel Code (noscript) */}
+      <noscript dangerouslySetInnerHTML={{ __html: `<img height="1" width="1" style="display:none"src="https://www.facebook.com/tr?id=487461332128996&ev=PageView&noscript=1"/>` }}/>
+      {/* // //End Meta Pixel Code (noscript) */}
+      {/* // // Google Tag Manager (noscript) */}
+      <noscript dangerouslySetInnerHTML={{ __html: `<iframe src="https://www.googletagmanager.com/ns.html?id=GTM-NGBCMHM" height="0"
+        width="0" style={{"display":"none","visibility":"hidden"}}></iframe>` }}/>
+      {/* // // End Google Tag Manager (noscript) */}
+  
       <title>{meta.title}</title>
     </Head>
     <HeaderFooterLayout breadcrumbs={false}>
